refactor: 로또 구매 금액은 음수가 될 수 없으므로 Int -> PositiveNumber로 변경
