+++ src/main/kotlin/lotto/domain/shop/LottoShop.kt
@@ -1,6 +1,8 @@
 package lotto.domain.shop
 
+import common.PositiveNumber
+
 interface LottoShop {
 
-    fun purchase(lottoPurchaseAmount: Int): LottoPurchaseResult
+    fun purchase(lottoPurchaseAmount: PositiveNumber): LottoPurchaseResult
 }
