+++ src/main/kotlin/lotto/domain/shop/RealLottoShop.kt
@@ -1,10 +1,12 @@
 package lotto.domain.shop
 
+import common.PositiveNumber
+
 class RealLottoShop : LottoShop {
 
-    override fun purchase(lottoPurchaseAmount: Int): LottoPurchaseResult {
+    override fun purchase(lottoPurchaseAmount: PositiveNumber): LottoPurchaseResult {
         return LottoPurchaseResult(
-            lottoGames = List(lottoPurchaseAmount / 1000) {
+            lottoGames = List(lottoPurchaseAmount.value / 1000) {
                 LottoGame(
                     numbers = (1..45).shuffled().take(6).sorted()
                 )
