+++ src/test/kotlin/lotto/domain/shop/RealLottoShopTest.kt
@@ -1,5 +1,6 @@
 package lotto.domain.shop
 
+import common.PositiveNumber
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 
@@ -8,13 +9,13 @@ class RealLottoShopTest : BehaviorSpec({
     Given("로또를 구매하면") {
 
         Then("구입 금액 만큼의 로또 게임 목록을 반환한다") {
-            val lottoGames = RealLottoShop().purchase(lottoPurchaseAmount = 5_000)
+            val lottoGames = RealLottoShop().purchase(PositiveNumber(5_000))
                 .lottoGames
             lottoGames.size shouldBe 5
         }
 
         Then("한장의 로또 게임에 담긴 로또 번호 개수는 6개이다") {
-            val lottoNumbers = RealLottoShop().purchase(lottoPurchaseAmount = 1_000)
+            val lottoNumbers = RealLottoShop().purchase(PositiveNumber(1_000))
                 .lottoGames
                 .first()
                 .numbers
@@ -22,7 +23,7 @@ class RealLottoShopTest : BehaviorSpec({
         }
 
         Then("로또 게임에 담긴 로또 번호 목록에는 중복이 없다") {
-            val lottoNumbers = RealLottoShop().purchase(lottoPurchaseAmount = 1_000)
+            val lottoNumbers = RealLottoShop().purchase(PositiveNumber(1_000))
                 .lottoGames
                 .first()
                 .numbers
@@ -30,7 +31,7 @@ class RealLottoShopTest : BehaviorSpec({
         }
 
         Then("로또 게임에 담긴 로또 번호 목록은 오름차순 정렬이 되어있다") {
-            val lottoNumbers = RealLottoShop().purchase(lottoPurchaseAmount = 1_000)
+            val lottoNumbers = RealLottoShop().purchase(PositiveNumber(1_000))
                 .lottoGames
                 .first()
                 .numbers
@@ -38,7 +39,7 @@ class RealLottoShopTest : BehaviorSpec({
         }
 
         Then("각 로또 번호의 범위는 0 이상 45 이하이다") {
-            val lottoNumbers = RealLottoShop().purchase(lottoPurchaseAmount = 1_000)
+            val lottoNumbers = RealLottoShop().purchase(PositiveNumber(1_000))
                 .lottoGames
                 .first()
                 .numbers
